【修改内容】： 1、修改OLED显示的逻辑 2、完善蓝牙控制履带车的代码 【影响范围】： 1、OLED模块 2、蓝牙模块 3、电机模块

+++ Hardware/Motor/Motor.h
@@ -10,6 +10,6 @@
 
 void Motor_GPIO_Init(void);
 void Motor_PWM_Init(void);
-void Motor_LEFT_Speed(int8_t Speed);
-void Motor_RIGHT_Speed(int8_t Speed);
+void Motor_LEFT_Speed(int32_t Speed);
+void Motor_RIGHT_Speed(int32_t Speed);
 #endif
